requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap==3.4.3
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
databases[postgresql]==0.8.0
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication & Security
python-decouple==3.8
cryptography==41.0.7

# HTTP & Web
httpx==0.25.2
jinja2==3.1.2
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
# mypy==1.7.1
